Clean up math functions
- Remove checking for standard functions from the build system
- Prefix all invocations with std::
- Replace suffixed functions with unprefixed versions
- Include <cmath> in all files that use it and remove the global include

e.g. floorf(x) -> std::floor(x)

+++ src/Attack.cpp
@@ -3,15 +3,16 @@
 #include "GameState.h"
 #include "RageUtil.h"
 #include "Song.h"
-
 #include "PlayerOptions.h"
 #include "PlayerState.h"
+
+#include <cmath>
 
 void Attack::GetAttackBeats( const Song *pSong, float &fStartBeat, float &fEndBeat ) const
 {
 	ASSERT( pSong != nullptr );
 	ASSERT_M( fStartSecond >= 0, ssprintf("StartSecond: %f",fStartSecond) );
-	
+
 	const TimingData &timing = pSong->m_SongTiming;
 	fStartBeat = timing.GetBeatFromElapsedTime( fStartSecond );
 	fEndBeat = timing.GetBeatFromElapsedTime( fStartSecond+fSecsRemaining );
@@ -34,13 +35,13 @@ void Attack::GetRealtimeAttackBeats( const Song *pSong, const PlayerState* pPlay
 
 	/* If reasonable, push the attack forward 8 beats so that notes on screen don't change suddenly. */
 	fStartBeat = std::min( GAMESTATE->m_Position.m_fSongBeat+8, pPlayerState->m_fLastDrawnBeat );
-	fStartBeat = truncf(fStartBeat)+1;
+	fStartBeat = std::trunc(fStartBeat) + 1;
 
 	const TimingData &timing = pSong->m_SongTiming;
 	const float lStartSecond = timing.GetElapsedTimeFromBeat( fStartBeat );
 	const float fEndSecond = lStartSecond + fSecsRemaining;
 	fEndBeat = timing.GetBeatFromElapsedTime( fEndSecond );
-	fEndBeat = truncf(fEndBeat)+1;
+	fEndBeat = std::trunc(fEndBeat) + 1;
 
 	// loading the course should have caught this.
 	ASSERT_M( fEndBeat >= fStartBeat, ssprintf("EndBeat %f >= StartBeat %f", fEndBeat, fStartBeat) );
@@ -49,7 +50,7 @@ void Attack::GetRealtimeAttackBeats( const Song *pSong, const PlayerState* pPlay
 bool Attack::operator== ( const Attack &rhs ) const
 {
 #define EQUAL(a) (a==rhs.a)
-	return 
+	return
 		EQUAL(level) &&
 		EQUAL(fStartSecond) &&
 		EQUAL(fSecsRemaining) &&
@@ -118,7 +119,7 @@ void AttackArray::UpdateStartTimes(float delta)
 /*
  * (c) 2003-2004 Chris Danford
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -128,7 +129,7 @@ void AttackArray::UpdateStartTimes(float delta)
  * copyright notice(s) and this permission notice appear in all copies of
  * the Software and that both the above copyright notice(s) and this
  * permission notice appear in supporting documentation.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
  * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT OF
