+++ src/arch/MemoryCard/MemoryCardDriverThreaded_Windows.cpp
@@ -6,6 +6,8 @@
 #include "PlayerNumber.h"
 #include "MemoryCardManager.h"
 
+#include <cmath>
+
 MemoryCardDriverThreaded_Windows::MemoryCardDriverThreaded_Windows()
 {
 	m_dwLastLogicalDrives = 0;
@@ -23,7 +25,7 @@ static bool TestReady( const RString &sDrive, RString &sVolumeLabelOut )
 	DWORD lpFileSystemFlags;
 	TCHAR szFileSystemNameBuffer[MAX_PATH];
 
-	if( !GetVolumeInformation( 
+	if( !GetVolumeInformation(
 		sDrive,
 		szVolumeNameBuffer,
 		sizeof(szVolumeNameBuffer),
@@ -117,7 +119,7 @@ void MemoryCardDriverThreaded_Windows::GetUSBStorageDevices( std::vector<UsbStor
 			continue;
 		}
 
-		// Testing hack:  Allow non-removable drive letters to be used if that 
+		// Testing hack:  Allow non-removable drive letters to be used if that
 		// driver letter is specified as a m_sMemoryCardOsMountPoint.
 
 		bool bIsSpecifiedMountPoint = false;
@@ -135,7 +137,7 @@ void MemoryCardDriverThreaded_Windows::GetUSBStorageDevices( std::vector<UsbStor
 			if( GetDriveType(sDrivePath) != DRIVE_REMOVABLE )
 			{
 				LOG->Trace( "not DRIVE_REMOVABLE" );
-				continue;	
+				continue;
 			}
 		}
 
@@ -143,7 +145,7 @@ void MemoryCardDriverThreaded_Windows::GetUSBStorageDevices( std::vector<UsbStor
 		if( !TestReady(sDrivePath, sVolumeLabel) )
 		{
 			LOG->Trace( "not TestReady" );
-			continue;	
+			continue;
 		}
 
 		vDevicesOut.push_back( UsbStorageDevice() );
@@ -172,7 +174,7 @@ void MemoryCardDriverThreaded_Windows::GetUSBStorageDevices( std::vector<UsbStor
 				&dwNumberOfFreeClusters,
 				&dwTotalNumberOfClusters ) )
 		{
-			usbd.iVolumeSizeMB = (int)roundf( dwTotalNumberOfClusters * (float)dwSectorsPerCluster * dwBytesPerSector / (1024*1024) );
+			usbd.iVolumeSizeMB = std::round( dwTotalNumberOfClusters * (float)dwSectorsPerCluster * dwBytesPerSector / (1024*1024) );
 		}
 	}
 }
@@ -209,7 +211,7 @@ void MemoryCardDriverThreaded_Windows::Unmount( UsbStorageDevice* pDevice )
 /*
  * (c) 2003-2004 Chris Danford
  * All rights reserved.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -219,7 +221,7 @@ void MemoryCardDriverThreaded_Windows::Unmount( UsbStorageDevice* pDevice )
  * copyright notice(s) and this permission notice appear in all copies of
  * the Software and that both the above copyright notice(s) and this
  * permission notice appear in supporting documentation.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
  * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT OF
